move Home component move react-scripts to dev dependency
