+++ package.json
@@ -14,7 +14,6 @@
     "react-dom": "^18.2.0",
     "react-icons": "^4.11.0",
     "react-router-dom": "^6.16.0",
-    "react-scripts": "5.0.1",
     "usehooks-ts": "^2.9.1"
   },
   "scripts": {
@@ -37,7 +36,8 @@
   },
   "devDependencies": {
     "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
-    "prettier": "3.0.3"
+    "prettier": "3.0.3",
+    "react-scripts": "5.0.1"
   },
   "proxy": "https://app.solarmoonanalytics.com"
 }
