+++ src/components/Home/Home.js
@@ -1,7 +1,8 @@
 import { useAuthenticator } from "@aws-amplify/ui-react";
 import { useEffect, useState } from "react";
-import Preloader from "./components/Preloader";
+import Preloader from "../Preloader";
 import { Card, CardHeader } from "react-bootstrap";
+import MetricTile from "./MetricTile";
 
 const Home = () => {
   const { user } = useAuthenticator((context) => [context.user]);
@@ -22,6 +23,7 @@ const Home = () => {
         <CardHeader>
           Welcome {user.attributes.email}!
           <br />
+          <MetricTile />
         </CardHeader>
       </Card>
       <Preloader load={load} />
